Fixes #36: Upgrade from <= 0.2.0 to 0.3.0 can result in no notification.
This is due to the fact that in 0.3.0 we started reading the
notification repo url with issue #25 so we can provide a more direct
url. That requires an OAuth token with the repo permission. Which we
don't have in versions 0.2.0 and lower.

If we fail to get a direct url to a notification and get a http status
code 401 or 404 we will ask for a new OAuth token.