Fixes #30: Notification should always be marked as read when clicked.
Currently the notification only gets marked as read when the user is
logged in when they click the notification on the mac.

Repro steps:

Log out of github in your default browser
Click a mac objective octocat notification on your mac
The notification will be removed from the mac, yet it will show up again
next time it checks for notifications since the notification is still
unread on github.